src/requirements.txt
aiXplain==0.2.11
Flask==3.0.3
langchain==0.1.16
langchain_google_genai==1.0.3
model_interfaces==0.0.1
python-dotenv==1.0.1
